% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CXtest.R
\name{CXtest}
\alias{CXtest}
\title{High dimensional MANOVA}
\usage{
CXtest(X.list, A = "Identity", pkg = "flare", biased = FALSE)
}
\arguments{
\item{X.list}{list of \eqn{g (\ge 2)} matrices from the samples.}

\item{A}{transformation matrix. \describe{
\item{"Identity"}{the \eqn{p\times p} identity matrix is used;}
\item{"Omega"}{the CLIME estimator (Cai, Liu and Luo, 2011) for precision matrix is used when the precision matrix is known to be sparse;}
\item{"Sigma"}{the adaptive thresholding estimator (Cai and Liu, 2011) for covariance  matrix is used when the precision matrix is not known to be sparse.}
}}

\item{pkg}{R package ("flare" or "fastclime") used for CLIME estimator when \code{A = "Omega"}.}

\item{biased}{a logical value indicating whether to use biased estimator for the pooled (co)variance.}
}
\value{
value of test statistic, p-value.
}
\description{
Testing the equality of multi-sample high dimensional mean vectors using the testing procedure by Cai and Xia (2014). For the two-sample test, see also Cai, Liu and Xia (2014).
}
\examples{
# generating a list of 3-sample data
set.seed(123)
X.list <- rdatalist(g = 3)
str(X.list)
CXtest(X.list)
CXtest(X.list, A = "Omega")

}
\references{
Cai, T., Liu, W., and Luo, X. (2011). A constrained \eqn{l_1} minimization approach to sparse precision matrix estimation. \emph{Journal of the American Statistical Association}, 106:594–607.

Cai, T. and Liu, W. (2011). Adaptive thresholding for sparse covariance matrix estimation. \emph{Journal of the American Statistical Association}, 106:672–684.

Cai, T., and Xia, Y. (2014). High-dimensional sparse MANOVA.  \emph{Journal of Multivariate Analysis}, 131:174–196.

Cai, T., Liu, W., and Xia, Y. (2014). Two-sample test of high dimensional means under dependence.  \emph{Journal of the Royal Statistical Society Series B: Statistical Methodology}, 76(2):349–372.
}
