% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HBtest.R
\name{HBtest}
\alias{HBtest}
\title{High dimensional MANOVA}
\usage{
HBtest(X.list, est = "LC")
}
\arguments{
\item{X.list}{list of \eqn{g (\ge 2)} matrices from the samples.}

\item{est}{method to be used for estimating tr(\eqn{\Sigma_i^2}). \describe{
\item{"BS"}{Bai and Saranadata, (1996);}
\item{"LC"}{Li and Chen, (2012);}
\item{"CQ"}{Chen and Qin, (2010).}
}}
}
\value{
value of test statistic and p-value.
}
\description{
Testing the equality of multi-sample high dimensional mean vectors using the testing procedure by Hu et. al. (2017) and Kong and Harrar (2021). For the two-sample test, see also Chen and Qin (2010).
}
\examples{
# generating a list of 3-sample data
set.seed(123)
X.list <- rdatalist(g = 3)
str(X.list)
HBtest(X.list)

}
\references{
Bai, Z. and Saranadasa, H. (1996). Effect of high dimension: By an example of a two sample problem. \emph{Statistica Sinica}, 6:311–329.

Li, J., Chen, S. (2012). Two sample tests for high-dimensional covariance matrices.  \emph{The Annals of Statistics}, 40(2):908–940.

Chen, S. and Qin, Y. (2010). A two-sample test for high-dimensional data with applications to gene-set testing.  \emph{The Annals of Statistics}, 38(2):808–835.

Hu, J., Bai, Z., Wang, C., and Wang, W. (2017). On testing the equality of high dimensional mean vectors with unequal covariance matrices.  \emph{Annals of the Institute of Statistical Mathematics}, 69(2):365–387.

Kong, X. and Harrar, S. (2021), High-dimensional MANOVA under weak conditions. \emph{Statistics: A Journal of Theoretical and Applied Statistics} 55:321–349.
}
