% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rdatalist.R
\name{rdatalist}
\alias{rdatalist}
\title{Simulate multi-sample of high-dimensional data for MANOVA.}
\usage{
rdatalist(
  g = 3,
  p = 100,
  nvec = 50,
  beta = 0,
  delta = 0,
  dist = "normal",
  err = "IND"
)
}
\arguments{
\item{g}{number of samples, \eqn{g \ge 2}.}

\item{p}{dimension.}

\item{nvec}{a vector of sample sizes, length \eqn{g}. If length 1, the same sample size will be used.}

\item{beta}{proportion of non-equal elements in mean \eqn{\mu_g}.}

\item{delta}{departure magnitudes in mean \eqn{\mu_g}.}

\item{dist}{innovation distribution. If length 1, the same distribution will be used. \describe{
\item{"normal"}{Normal(mean = 0, sd = 1);}
\item{"t"}{t(df = 3);}
\item{"cauchy"}{Cauchy(location = 0, scale = 0.1);}
\item{"gamma"}{centered Gamma(shape = 4, scale = 2).}
}}

\item{err}{dependence structures for error processes \eqn{\{\varepsilon_{ijk}: k = 1,\ldots, p\}}. If length 1, the same dependence structure will be used. \describe{
\item{"IND"}{(Independent) independent generation from the given innovation distribution.}
\item{"WD"}{(Weakly Dependent) random generation of ARMA(2, 2) model from the given innovation distribution, where autoregressive coefficients \eqn{\phi = (0.4, −0.1)} and moving-average coefficients \eqn{\theta = (0.2, 0.3)};}
\item{"SD"}{(Strongly Dependent) random generation of AR(1) model from the given innovation distribution, where autoregressive coefficients \eqn{\phi = 0.9};}
\item{"LD"}{(Long-range Dependent) \eqn{\varepsilon_{ij}=U^t\eta_{ij}}, where \eqn{\eta_{ij}} is independently generated \eqn{p}-dim vector from the given innovation distribution and \eqn{U^tU = A} is the Cholesky factorization of the \eqn{p \times p} matrix \eqn{A=(a_{st})}, \eqn{a_{st}=0.5[(|s-t|+1)^{1.4}+(|s-t|-1)^{1.4}-2|s-t|^{1.4}]}.}
}}
}
\value{
a list of \eqn{n_i \times p} sample matrices for \eqn{i = 1,\ldots, g}.
}
\description{
The data are generated from various time series dependence structures and innovation distributions:
\deqn{X_{ij}=\mu_{i}+\varepsilon_{ij},\quad 1\le i \le g,\ 1\le j\le n_i.}
}
\examples{
set.seed(123)
X.list <- rdatalist()
str(X.list)
X.list <- rdatalist(g = 3,
                    p = 300,
                    nvec = c(80, 90, 100),
                    beta = 0.1,
                    delta = 0.5,
                    dist = c("t", "normal","normal"),
                    err = c("IND","WD","SD")
                    )
str(X.list)

}
