% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SKtest.R
\name{SKtest}
\alias{SKtest}
\title{High dimensional MANOVA}
\usage{
SKtest(X.list)
}
\arguments{
\item{X.list}{list of \eqn{g (\ge 2)} matrices from the samples.}
}
\value{
value of test statistic and p-value.
}
\description{
Testing the equality of multi-sample high dimensional mean vectors using the testing procedure by Srivastava and Kubokawa (2013). For the two-sample test, see also Srivastava and Du (2008).
}
\examples{
# generating a list of 3-sample data
set.seed(123)
X.list <- rdatalist(g = 3)
str(X.list)
SKtest(X.list)

}
\references{
Srivastava, M. and Du, M. (2008). A test for the mean vector with fewer observations than the dimension.  \emph{Journal of Multivariate Analysis}, 99:386–402.

Srivastava, M. and Kubokawa, T. (2013). Tests for multivariate analysis of variance in high dimension under non-normality.  \emph{Journal of Multivariate Analysis}, 115:204–216.
}
