% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KVtest.R
\name{KVtest}
\alias{KVtest}
\title{High dimensional MANOVA}
\usage{
KVtest(X.list, method = "GCT", weight.fun = "parz", L = 10, npower = 0)
}
\arguments{
\item{X.list}{list of \eqn{g (\ge 2)} matrices from the samples.}

\item{method}{process to be used for estimating scale parameter.\describe{
\item{"GCT"}{generalized component test (Gregory et. al., 2015);}
\item{"MPT"}{more powerful test (Zhang and Wang, 2021).}
}}

\item{weight.fun}{window weight function, "parz" for Parzen weight and "trapez" for trapezoid weight.}

\item{L}{value of the window width. If \code{NA}, the elbow method is used for searching it.}

\item{npower}{power (negative) 0, 1, or 2 of n at which the center is estimated.}
}
\value{
value of test statistic, p-value and window width.
}
\description{
Testing the equality of multi-sample high dimensional mean vectors using the testing procedure by Kong, et. al. (Submitted). For the two-sample test, see also Gregory et. al (2015) or Zhang and Wang (2021).
}
\examples{
# generating a list of 3-sample data
set.seed(123)
X.list <- rdatalist(g = 3)
str(X.list)
KVtest(X.list)
KVtest(X.list, method = "MPT", L = NA)

}
\references{
Gregory, K., Carroll, R., Baladandayuthapani, V., and Lahiri, S. (2015). A two-sample test for equality of means in high dimension.  \emph{Journal of the American Statistical Association}, 110(510):837–849.

Zhang, H. and Wang, H. (2021). A more powerful test of equality of high-dimensional two-sample means.  \emph{Computational Statistics & Data Analysis}, 164:107318.

Kong, X., Villasante-Tezanos, A., Fardo, D.. and Harrar, S. (submitted). Generalized composite multi-sample tests for high-dimensional data. \emph{}
}
